requirements.txt
streamlit
beautifulsoup4
requests
pandas
lxml
openai
playwright
python-dotenv
matplotlib